Move constant to common
